Major refactor to remove pokenav ewram macros
