+++ asm/pokenav.s
@@ -5,248 +5,6 @@
 
 	.text
 
-	thumb_func_start sub_80F6074
-sub_80F6074: @ 80F6074
-	push {r4-r7,lr}
-	ldr r1, _080F6120 @ =gPokenavStructPtr
-	ldr r6, [r1]
-	ldr r1, _080F6124 @ =0x00008770
-	adds r3, r6, r1
-	ldrh r1, [r3]
-	lsls r0, 16
-	asrs r7, r0, 16
-	adds r2, r7, r1
-	strh r2, [r3]
-	ldr r5, _080F6128 @ =0x00008774
-	adds r4, r6, r5
-	lsls r0, r2, 16
-	asrs r0, 16
-	ldrh r1, [r4]
-	mov r12, r1
-	movs r5, 0
-	ldrsh r1, [r4, r5]
-	cmp r0, r1
-	ble _080F60A4
-	subs r0, r2, 0x1
-	mov r1, r12
-	subs r0, r1
-	strh r0, [r3]
-_080F60A4:
-	ldrh r1, [r3]
-	movs r2, 0
-	ldrsh r0, [r3, r2]
-	cmp r0, 0
-	bge _080F60B6
-	adds r0, r1, 0x1
-	ldrh r5, [r4]
-	adds r0, r5
-	strh r0, [r3]
-_080F60B6:
-	ldr r0, _080F612C @ =0x00008772
-	adds r3, r6, r0
-	ldrh r0, [r3]
-	adds r2, r7, r0
-	strh r2, [r3]
-	lsls r0, r2, 16
-	asrs r0, 16
-	ldrh r1, [r4]
-	mov r12, r1
-	movs r5, 0
-	ldrsh r1, [r4, r5]
-	cmp r0, r1
-	ble _080F60D8
-	subs r0, r2, 0x1
-	mov r1, r12
-	subs r0, r1
-	strh r0, [r3]
-_080F60D8:
-	ldrh r1, [r3]
-	movs r2, 0
-	ldrsh r0, [r3, r2]
-	cmp r0, 0
-	bge _080F60EA
-	adds r0, r1, 0x1
-	ldrh r5, [r4]
-	adds r0, r5
-	strh r0, [r3]
-_080F60EA:
-	ldr r0, _080F6130 @ =0x0000876e
-	adds r3, r6, r0
-	ldrh r0, [r3]
-	adds r2, r7, r0
-	strh r2, [r3]
-	lsls r0, r2, 16
-	asrs r0, 16
-	ldrh r5, [r4]
-	movs r6, 0
-	ldrsh r1, [r4, r6]
-	cmp r0, r1
-	ble _080F6108
-	subs r0, r2, 0x1
-	subs r0, r5
-	strh r0, [r3]
-_080F6108:
-	ldrh r1, [r3]
-	movs r2, 0
-	ldrsh r0, [r3, r2]
-	cmp r0, 0
-	bge _080F611A
-	adds r0, r1, 0x1
-	ldrh r4, [r4]
-	adds r0, r4
-	strh r0, [r3]
-_080F611A:
-	pop {r4-r7}
-	pop {r0}
-	bx r0
-	.align 2, 0
-_080F6120: .4byte gPokenavStructPtr
-_080F6124: .4byte 0x00008770
-_080F6128: .4byte 0x00008774
-_080F612C: .4byte 0x00008772
-_080F6130: .4byte 0x0000876e
-	thumb_func_end sub_80F6074
-
-	thumb_func_start sub_80F6134
-sub_80F6134: @ 80F6134
-	push {r4-r6,lr}
-	ldr r0, _080F6194 @ =gPokenavStructPtr
-	ldr r3, [r0]
-	ldr r1, _080F6198 @ =0x000087c9
-	adds r0, r3, r1
-	ldrb r0, [r0]
-	cmp r0, 0
-	beq _080F61E0
-	ldr r2, _080F619C @ =0x000087dc
-	adds r5, r3, r2
-	movs r4, 0
-	ldrsh r1, [r5, r4]
-	ldr r6, _080F61A0 @ =0x00008774
-	adds r4, r3, r6
-	movs r2, 0
-	ldrsh r0, [r4, r2]
-	subs r0, 0x7
-	cmp r1, r0
-	bge _080F61AC
-	subs r6, 0x8
-	adds r1, r3, r6
-	movs r0, 0
-	strh r0, [r1]
-	ldrh r0, [r5]
-	ldr r1, _080F61A4 @ =0x00008770
-	adds r2, r3, r1
-	strh r0, [r2]
-	ldrh r1, [r5]
-	ldr r5, _080F61A8 @ =0x0000876e
-	adds r0, r3, r5
-	strh r1, [r0]
-	ldrh r5, [r2]
-	adds r0, r5, 0x7
-	adds r6, 0x6
-	adds r2, r3, r6
-	strh r0, [r2]
-	lsls r0, 16
-	asrs r0, 16
-	ldrh r3, [r4]
-	movs r6, 0
-	ldrsh r1, [r4, r6]
-	cmp r0, r1
-	ble _080F61F4
-	adds r0, r5, 0x6
-	subs r0, r3
-	strh r0, [r2]
-	b _080F61F4
-	.align 2, 0
-_080F6194: .4byte gPokenavStructPtr
-_080F6198: .4byte 0x000087c9
-_080F619C: .4byte 0x000087dc
-_080F61A0: .4byte 0x00008774
-_080F61A4: .4byte 0x00008770
-_080F61A8: .4byte 0x0000876e
-_080F61AC:
-	ldrh r0, [r4]
-	subs r0, 0x7
-	ldr r2, _080F61D4 @ =0x00008770
-	adds r1, r3, r2
-	strh r0, [r1]
-	ldrh r1, [r4]
-	ldr r6, _080F61D8 @ =0x00008772
-	adds r0, r3, r6
-	strh r1, [r0]
-	ldrh r1, [r5]
-	subs r2, 0x2
-	adds r0, r3, r2
-	strh r1, [r0]
-	ldrh r0, [r4]
-	subs r0, r1
-	movs r1, 0x7
-	subs r1, r0
-	ldr r4, _080F61DC @ =0x0000876c
-	adds r0, r3, r4
-	b _080F61F2
-	.align 2, 0
-_080F61D4: .4byte 0x00008770
-_080F61D8: .4byte 0x00008772
-_080F61DC: .4byte 0x0000876c
-_080F61E0:
-	ldr r5, _080F61FC @ =0x000087dc
-	adds r2, r3, r5
-	ldrh r1, [r2]
-	ldr r6, _080F6200 @ =0x0000876c
-	adds r0, r3, r6
-	strh r1, [r0]
-	ldrh r1, [r2]
-	ldr r2, _080F6204 @ =0x0000876e
-	adds r0, r3, r2
-_080F61F2:
-	strh r1, [r0]
-_080F61F4:
-	pop {r4-r6}
-	pop {r0}
-	bx r0
-	.align 2, 0
-_080F61FC: .4byte 0x000087dc
-_080F6200: .4byte 0x0000876c
-_080F6204: .4byte 0x0000876e
-	thumb_func_end sub_80F6134
-
-	thumb_func_start sub_80F6208
-sub_80F6208: @ 80F6208
-	push {lr}
-	ldr r0, _080F623C @ =gPokenavStructPtr
-	ldr r1, [r0]
-	ldr r2, _080F6240 @ =0x00008fe6
-	adds r0, r1, r2
-	movs r2, 0
-	strb r2, [r0]
-	ldr r3, _080F6244 @ =0x00008fe7
-	adds r0, r1, r3
-	strb r2, [r0]
-	ldr r0, _080F6248 @ =0x00008fe8
-	adds r2, r1, r0
-	movs r0, 0xFF
-	strb r0, [r2]
-	ldr r2, _080F624C @ =0x00006dac
-	adds r1, r2
-	ldrb r0, [r1]
-	cmp r0, 0
-	bne _080F6238
-_080F622E:
-	bl sub_80F6250
-	lsls r0, 24
-	cmp r0, 0
-	bne _080F622E
-_080F6238:
-	pop {r0}
-	bx r0
-	.align 2, 0
-_080F623C: .4byte gPokenavStructPtr
-_080F6240: .4byte 0x00008fe6
-_080F6244: .4byte 0x00008fe7
-_080F6248: .4byte 0x00008fe8
-_080F624C: .4byte 0x00006dac
-	thumb_func_end sub_80F6208
-
 	thumb_func_start sub_80F6250
 sub_80F6250: @ 80F6250
 	push {r4-r7,lr}
