+++ include/pokenav.h
@@ -63,7 +63,8 @@ struct UnknownStruct_Pokenav_8FF0 {
 
 struct UnkPokenavStruct {
     /*0x0000*/ u16 unk0[6][0x40];
-    /*0x0300*/ u8 filler0300[0x6];
+    /*0x0300*/ void (*unk300)(void);
+    /*0x0304*/ u16 unk304;
     /*0x0306*/ u16 unk306;
     /*0x0306*/ int unk308;
     /*0x030C*/ s16 unk030C;
@@ -97,12 +98,23 @@ struct UnkPokenavStruct {
     /*0x6DA0*/ u16 unk6DA0;
     /*0x6DA2*/ u16 unk6DA2;
     /*0x6DA4*/ u16 taskId6DA4;
-    /*0x6DA6*/ u8 filler6DA6[0x6];
+    /*0x6DA6*/ u8 filler6DA6[0x5];
+    /*0x6DAB*/ u8 unk6DAB;
     /*0x6DAC*/ u8 unk6DAC;
     /*0x6DAD*/ s8 unk6DAD;
     /*0x6DAE*/ s8 unk6DAE;
     /*0x6DAF*/ u8 filler6DAF[0x3];
-    /*0x6DB2*/ u8 unk6DB2[0x66];
+    /*0x6DB2*/ u8 unk6DB2[0x2A];
+    /*0x6DDC*/ s8 unk6DDC;
+    /*0x6DDD*/ u8 filler6DDD[0x13];
+    /*0x6DF0*/ s8 unk6DF0;
+    /*0x6DF1*/ u8 filler6DF1[0x0B];
+    /*0x6DFC*/ u8 unk6DFC;
+    /*0x6DFD*/ u8 filler6DFD[0x17];
+    /*0x6E14*/ u8 unk6E14;
+    /*0x6E15*/ u8 unk6E15;
+    /*0x6E16*/ u8 unk6E16;
+    /*0x6E17*/ u8 unk6E17;
     /*0x6E18*/ struct RegionMap regionMap;
     /*0x7698*/ u16 unk7698;
     /*0x769A*/ u16 unk769A;
@@ -154,7 +166,7 @@ struct UnkPokenavStruct {
     /*0x8FE4*/ u16 unk8FE4;
     /*0x8FE6*/ u8 unk8FE6;
     /*0x8FE7*/ u8 unk8FE7;
-    /*0x8FE8*/ u8 unk8FE8;
+    /*0x8FE8*/ s8 unk8FE8;
     /*0x8fe9*/ s8 unk8fe9;
     /*0x8fea*/ s8 unk8FEA;
     /*0x8feb*/ s8 unk8FEB;
@@ -256,5 +268,206 @@ void sub_80F105C(void);
 void sub_80F43D4(u8 *);
 void sub_80F063C(s16 arg0);
 void sub_80F6074(s16 arg0);
+void sub_80F6208(void);
+bool8 sub_80F6250();
+
+void sub_80EBCA8();
+void sub_80EEE20();
+bool8 sub_80EEE54();
+void sub_80EEE08();
+void sub_80EED2C(u8);
+void sub_80EC268();
+void sub_80EED1C();
+void sub_80EE9C0(u8, u8, u8);
+bool8 sub_80EEA0C();
+bool8 sub_80EEC10();
+void sub_80EED9C();
+void sub_80EDB88();
+void sub_80EC4A0();
+void sub_80EC81C();
+void sub_80EE96C();
+void sub_80EE3D8();
+bool8 sub_80EEF34();
+void sub_80EED0C();
+void sub_80EC67C();
+void sub_80EC86C();
+bool8 sub_80EEC90();
+void sub_80ED620();
+void sub_80EC960();
+void sub_80ED01C();
+void sub_80ECC08();
+void sub_80ED31C();
+void sub_80ED4D8();
+void sub_80ED858();
+void sub_80EDDBC();
+void sub_80EDE70();
+void sub_80EDEE4();
+void sub_80EE06C();
+void sub_80EE294();
+void sub_80EE58C();
+void sub_80EE658();
+void sub_80EE8F4();
+void sub_80EEDC4();
+void ShowMapNamePopUpWindow(void);
+void sub_80F0954(u16, u16, u16);
+bool8 sub_80F098C(void);
+u8 *sub_80F445C(u8*, u16);
+void sub_80F081C(u8);
+bool8 sub_80F0944(void);
+u8 *sub_80F4428(u8*, u16, u8);
+void sub_80F700C(u8*, u16);
+void sub_80F0B24(void);
+bool8 sub_80F0B44(void);
+void sub_80F0C28(void);
+bool8 sub_80F0C48(void);
+void LoadTrainerEyesDescriptionLines(void);
+bool8 sub_80F0D5C(void);
+void sub_80F0EC0(void);
+bool8 sub_80F0EF4(void);
+void sub_80F0F64(void);
+void sub_80F0FA0(void);
+bool8 sub_80F0FEC(void);
+bool8 sub_80F0718(void);
+void sub_80F0FFC(u8);
+void sub_80F19DC(u8*);
+bool8 sub_80F1080(void);
+void sub_80F1614(void);
+void DrawMonRibbonIcons(void);
+void sub_80F13FC(void);
+void sub_80F1438(void);
+void sub_80F1494(void);
+bool8 sub_80F162C(u8);
+void sub_80F01E0(u16);
+void sub_80F19FC(void);
+void sub_80F1A74(void);
+void sub_80F1A80(void);
+void sub_80F1A90(void);
+bool8 sub_80F1AC4(void);
+void sub_80F1B8C(u8);
+bool8 sub_80F1BC8(u8);
+void sub_80F2458(u8);
+void sub_80F2514(u8);
+void sub_80F1DF0(void);
+void sub_80F2218(struct Sprite *sprite);
+bool8 sub_80F1E50(void);
+bool8 sub_80F22F8(void);
+void sub_80F2108(void);
+bool8 sub_80F1E6C(void);
+void sub_80F1E84(void);
+void sub_80F2148(void);
+void sub_80F2240(struct Sprite *sprite);
+bool8 sub_80F1F10(void);
+bool8 sub_80F2360(void);
+void sub_80F2170(void);
+bool8 sub_80F23C8(void);
+void sub_80F21F8(void);
+void sub_80F1FF0(void);
+void sub_80F208C(void);
+void sub_80F22B0(struct Sprite *sprite);
+void sub_80F240C(struct Sprite *sprite);
+void sub_80F2598(void);
+void sub_80F2620(void);
+bool8 sub_80F26BC(void);
+void sub_80F2C58(struct Sprite *sprite);
+void sub_80F2D04(u8);
+void sub_80F2D6C(u8);
+void sub_80F2DD8(void);
+void sub_80F2DF4(void);
+void sub_80F2FEC(struct Sprite *sprite);
+void sub_80F2FB0(void);
+void sub_80F3008(u8);
+void sub_80F3130(void);
+void sub_80F3264(void);
+void sub_80F3294(u8);
+void sub_80F35B4(void);
+void sub_80F363C(struct Sprite *sprite);
+void SetMonMarkings(u16, u16, u8);
+void sub_80F36F0(void);
+bool8 sub_80F3724(void);
+void sub_80F379C(void);
+bool8 sub_80F37D0(void);
+void sub_80F38B8(void);
+bool8 sub_80F38EC(void);
+void sub_80F3970(void);
+bool8 sub_80F39A4(void);
+void sub_80F3B00(void);
+bool8 sub_80F3B58(void);
+void sub_80F3B94(void);
+bool8 sub_80F3BD4(void);
+void sub_80F3C2C(void);
+void sub_80F3F20(u8, u8);
+void sub_80F3FAC(void);
+void sub_80F3FF0(void);
+bool8 sub_80F4024(void);
+void sub_80F42C4(u8*);
+void sub_80F4394(void);
+u8 *sub_80F6514(u8*, u16, u8);
+u8 *sub_80F443C(u8 *, u16);
+void sub_80F55AC(u8*, struct UnkPokenav11 *);
+void sub_80F4CF0(void);
+void sub_80F4D44(void);
+bool8 sub_80F4D88(void);
+void sub_80F53EC(struct UnkPokenav11*, struct UnkPokenav11*);
+bool8 sub_80F5504(void);
+bool8 sub_80F5264(void);
+bool8 sub_80F52F8(void);
+bool8 sub_80F5364(void);
+
+void sub_80F0900(void);
+void sub_80F01A4(void);
+void sub_80EFD3C(void);
+void sub_8095C8C();
+void sub_80EFDA0(void);
+void sub_80EFD74(void);
+bool8 sub_80EFC64(void);
+void sub_80EFC3C(void);
+void sub_80EF624(const u16 *, const u16 *, u8, u8, u16 *);
+void sub_80EF7D4(void);
+void sub_80EF54C(u8);
+void sub_80EF58C(u8);
+void sub_80F6FFC();
+void sub_80F6FB8(bool8);
+void sub_80F6DB8();
+bool8 sub_80F6E9C();
+bool8 sub_80F6ED4();
+bool8 sub_80F70FC();
+void sub_80F708C(s8);
+void sub_80F6F10();
+void sub_80F15A8(void);
+void sub_80F6A4C();
+bool8 sub_80F6AF0();
+u8 sub_80F68E8();
+void sub_80F66E0();
+void sub_80F638C();
+bool8 sub_80F63D0();
+void sub_80EFF34();
+bool8 sub_80EFF68();
+void sub_80F6134();
+u8 sub_80F5DD4();
+void sub_80F0264(u8);
+bool8 sub_80F02A0();
+void sub_80EF9F8(void);
+bool8 sub_80EFBDC(bool8);
+void sub_80EFBB0(void);
+void sub_80EEFBC(u8);
+void sub_80EF814(void);
+void sub_80EF840(void);
+bool8 sub_80EF874(void);
+bool8 sub_80F6250();
+void sub_80F6C20();
+void sub_80EF248(u8);
+bool8 sub_80EF284(u8);
+void sub_80EF428(u8, u8);
+bool8 sub_80EEF78();
+void sub_80EBC10();
+void sub_80EBDBC(void (*func)(void));
+void sub_80EBBE8();
+void sub_80EBDD8();
+void sub_80EBD90();
+void sub_80EBD18();
+void sub_80FB260();
+void sub_80EFE7C(void);
+void sub_80F5BF0();
+void sub_80F6F64();
 
 #endif // GUARD_POKENAV_H
