+++ include/region_map.h
@@ -15,46 +15,46 @@ struct RegionMap
     /*0x16*/ u8 unk16;
     /*0x17*/ u8 everGrandeCityArea;
     /*0x18*/ u8 (*inputCallback)(void);
-    struct Sprite *cursorSprite;
-    struct Sprite *playerIconSprite;
-    s32 bg2x;
-    s32 bg2y;
-    u32 unk2C;  //
-    u32 unk30;  //  Map rotation parameters
-    u32 unk34;  //  Likely a scrapped feature
-    u32 unk38;  //
-    s32 unk3C;
-    s32 unk40;
-    s32 unk44;
-    s32 unk48;
-    s32 unk4C;
-    s32 unk50;
-    u16 cursorPosX;
-    u16 cursorPosY;
-    u16 cursorTileTag;
-    u16 cursorPaletteTag;
-    s16 scrollX;
-    s16 scrollY;
-    s16 unk60;
-    s16 unk62;
-    u16 unk64;
-    u16 unk66;
-    u16 unk68;
-    u16 unk6A;
-    u16 unk6C;
-    u16 unk6E;
-    u16 playerIconTileTag;
-    u16 playerIconPaletteTag;
-    u16 unk74;
-    u16 unk76;
-    bool8 zoomed;
-    u8 initStep;
-    s8 unk7A;
-    s8 cursorDeltaX;
-    s8 cursorDeltaY;
-    bool8 needUpdateVideoRegs;
-    bool8 blinkPlayerIcon;
-    bool8 playerIsInCave;
+    /*0x1C*/ struct Sprite *cursorSprite;
+    /*0x20*/ struct Sprite *playerIconSprite;
+    /*0x24*/ s32 bg2x;
+    /*0x28*/ s32 bg2y;
+    /*0x2C*/ u32 unk2C;  //
+    /*0x30*/ u32 unk30;  //  Map rotation parameters
+    /*0x34*/ u32 unk34;  //  Likely a scrapped feature
+    /*0x38*/ u32 unk38;  //
+    /*0x3C*/ s32 unk3C;
+    /*0x40*/ s32 unk40;
+    /*0x44*/ s32 unk44;
+    /*0x48*/ s32 unk48;
+    /*0x4C*/ s32 unk4C;
+    /*0x50*/ s32 unk50;
+    /*0x54*/ u16 cursorPosX;
+    /*0x56*/ u16 cursorPosY;
+    /*0x58*/ u16 cursorTileTag;
+    /*0x5A*/ u16 cursorPaletteTag;
+    /*0x5C*/ s16 scrollX;
+    /*0x5E*/ s16 scrollY;
+    /*0x60*/ s16 unk60;
+    /*0x62*/ s16 unk62;
+    /*0x64*/ u16 unk64;
+    /*0x66*/ u16 unk66;
+    /*0x68*/ u16 unk68;
+    /*0x6A*/ u16 unk6A;
+    /*0x6C*/ u16 unk6C;
+    /*0x6E*/ u16 unk6E;
+    /*0x70*/ u16 playerIconTileTag;
+    /*0x72*/ u16 playerIconPaletteTag;
+    /*0x74*/ u16 unk74;
+    /*0x76*/ u16 unk76;
+    /*0x78*/ bool8 zoomed;
+    /*0x79*/ u8 initStep;
+    /*0x7A*/ s8 unk7A;
+    /*0x7B*/ s8 cursorDeltaX;
+    /*0x7C*/ s8 cursorDeltaY;
+    /*0x7D*/ bool8 needUpdateVideoRegs;
+    /*0x7E*/ bool8 blinkPlayerIcon;
+    /*0x7F*/ bool8 playerIsInCave;
     /*0x080*/ u8 filler80[0x100];
     /*0x180*/ u8 cursorSmallImage[0x100];
     /*0x280*/ u8 cursorLargeImage[0x600];
