+++ ld_script.txt
@@ -296,7 +296,6 @@ SECTIONS {
         src/easy_chat_2.o(.text);
         src/pokenav_before.o(.text);
         src/pokenav.o(.text);
-        src/menu_specialized.o(.text);
         asm/pokenav.o(.text);
         src/trainers_eye.o(.text);
         src/mon_markings.o(.text);
@@ -697,6 +696,7 @@ SECTIONS {
         src/easy_chat_1.o(.rodata);
         src/easy_chat_2.o(.rodata);
         src/easy_chat_2.o(.data);
+        src/pokenav_before.o(.rodata);
         src/pokenav.o(.rodata);
         data/pokenav.o(.rodata);
         src/trainers_eye.o(.rodata);
